requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Cloud Storage (DigitalOcean Spaces / S3)
boto3==1.34.34

# PDF Processing
pdf2image==1.17.0
Pillow==10.2.0

# OCR
pytesseract==0.3.10

# Google Gemini Integration
google-generativeai==0.3.2

# Environment Variables
python-dotenv==1.0.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1

# Background Tasks
celery[redis]==5.3.6
redis==5.0.1

# Testing
pytest==8.0.0
httpx==0.26.0
